chore: restructure test cases
