+++ test/basic.test.ts
@@ -1,6 +1,5 @@
 import { describe, test, expect } from 'vitest'
 
-import run, { Options } from '../src/run'
 import {
   CharType,
   checkCharType,
@@ -15,17 +14,6 @@ import {
 import join from '../src/join'
 import processRule from '../src/process-rule'
 import findIgnoredMarks from '../src/ignore'
-
-import spacePunctuation from '../src/legacy-rules/space-punctuation'
-import spaceBrackets from '../src/legacy-rules/space-brackets'
-import spaceQuotes from '../src/legacy-rules/space-quotes'
-import spaceFullWidthContent from '../src/legacy-rules/space-full-width-content'
-import unifyPunctuation from '../src/legacy-rules/unify-punctuation'
-import caseTraditional from '../src/legacy-rules/case-traditional'
-import caseDatetime from '../src/legacy-rules/case-datetime'
-import caseDatetimeZh from '../src/legacy-rules/case-datetime-zh'
-
-const lint = (...args) => run(...(args as [string, Options])).result
 
 const purify = (arr) =>
   arr.map((item) => (Array.isArray(item) ? purify(item) : item))
@@ -774,178 +762,5 @@ describe('process rules', () => {
       }
     })
     expect(join(data.tokens)).toBe(`关注（watch）你关心的仓库。`)
-  })
-})
-
-describe('lint by rule', () => {
-  test('space between half-width content and full-width content', () => {
-    expect(
-      lint('汉字和English之间需要有空格比如 half width content。', {
-        rules: [spaceFullWidthContent]
-      })
-    ).toBe('汉字和 English 之间需要有空格比如 half width content。')
-  })
-  test('space beside brackets', () => {
-    const rules = [spaceFullWidthContent, spaceBrackets]
-    expect(
-      lint('汉字和Eng(lish之间)需要有空格比如 half width content。', { rules })
-    ).toBe('汉字和 Eng(lish 之间) 需要有空格比如 half width content。')
-    expect(
-      lint('汉字和Eng（lish之间）需要有空格比如 half width content。', {
-        rules
-      })
-    ).toBe('汉字和 Eng（lish 之间）需要有空格比如 half width content。')
-    expect(
-      lint(
-        '汉 (字 ) 和Eng（lish之间）需（ 要）有(空格)比如 half w(i)dth content。',
-        { rules }
-      )
-    ).toBe(
-      '汉 (字) 和 Eng（lish 之间）需（要）有 (空格) 比如 half w(i)dth content。'
-    )
-  })
-  test('unifies full-width/half-width mixed punctuation usage', () => {
-    const rules = [spaceFullWidthContent, unifyPunctuation]
-    expect(
-      lint('汉字和English之间需要有空格比如 half width content.', { rules })
-    ).toBe('汉字和 English 之间需要有空格比如 half width content。')
-    expect(
-      lint('汉字和"English"之间需要有空格比如 half width content.', { rules })
-    ).toBe('汉字和“English”之间需要有空格比如 half width content。')
-    expect(
-      lint('汉字和English之间需要:有；空格比如 half width content.', { rules })
-    ).toBe('汉字和 English 之间需要：有；空格比如 half width content。')
-  })
-  test('space beside punctuations', () => {
-    expect(
-      lint(
-        '汉字和Engl,is。h之间,需，要有, 空， 格 ，比 , 如 half width content.',
-        { rules: [spacePunctuation] }
-      )
-    ).toBe('汉字和Engl,is。h之间, 需，要有, 空，格，比, 如 half width content.')
-  })
-  test('space beside quotes', () => {
-    expect(
-      lint(`汉"字'和'English之间"需“要‘有’空”格比如 h'a'lf "width" content.`, {
-        rules: [spaceQuotes]
-      })
-    ).toBe(
-      `汉 "字 '和' English之间" 需“要‘有’空”格比如 h'a'lf "width" content.`
-    )
-  })
-  test('traditional characters', () => {
-    expect(
-      lint(
-        `老師說：「你們要記住國父說的『青年要立志做大事，不要做大官』這句話。」`,
-        { rules: [caseTraditional] }
-      )
-    ).toBe(`老師說：“你們要記住國父說的‘青年要立志做大事，不要做大官’這句話。”`)
-    expect(
-      lint(
-        `孔子曰：「求，周任有言曰：『陳力就列，不能則止。』危而不持，顛而不扶，則將焉用彼相矣？」`,
-        { rules: [caseTraditional] }
-      )
-    ).toBe(
-      `孔子曰：“求，周任有言曰：‘陳力就列，不能則止。’危而不持，顛而不扶，則將焉用彼相矣？”`
-    )
-    expect(
-      lint(
-        `我們雖不敢希望每個人能有范文正公「先天下之憂而憂，後天下之樂而樂」的大志向，但至少要有陶侃勤懇不懈的精神`,
-        { rules: [caseTraditional] }
-      )
-    ).toBe(
-      `我們雖不敢希望每個人能有范文正公“先天下之憂而憂，後天下之樂而樂”的大志向，但至少要有陶侃勤懇不懈的精神`
-    )
-    expect(
-      lint(`所謂忠恕，也就是「盡己之心，推己及人」的意思。`, {
-        rules: [caseTraditional]
-      })
-    ).toBe(`所謂忠恕，也就是“盡己之心，推己及人”的意思。`)
-  })
-  test('datetime', () => {
-    const rules = [
-      spaceFullWidthContent,
-      spacePunctuation,
-      caseDatetime,
-      caseDatetimeZh
-    ]
-    expect(lint('2019年06月26号 2019-06-26 12:00', { rules })).toBe(
-      '2019年06月26号 2019-06-26 12:00'
-    )
-  })
-  test('abbr', () => {
-    expect(lint('运行时 + 编译器 vs. 只包含运行时')).toBe(
-      '运行时 + 编译器 vs. 只包含运行时'
-    )
-  })
-})
-
-describe('lint special/edge cases', () => {
-  test('URL', () => {
-    expect(lint('Vue.js 是什么')).toBe('Vue.js 是什么')
-    expect(lint('www.vuejs.org')).toBe('www.vuejs.org')
-    expect(lint('https://vuejs.org')).toBe('https://vuejs.org')
-  })
-  test('slash character', () => {
-    expect(lint('想知道 Vue 与其它库/框架有哪些区别')).toBe(
-      '想知道 Vue 与其它库/框架有哪些区别'
-    )
-  })
-  test('special characters', () => {
-    expect(lint('Vue (读音 /vjuː/，类似于)')).toBe('Vue (读音 /vjuː/，类似于)')
-  })
-  test('plural brackets', () => {
-    expect(lint('3 minite(s) left')).toBe('3 minite(s) left')
-  })
-  test('single quote for shorthand', () => {
-    expect(lint(`how many user's here`)).toBe(`how many user's here`)
-    expect(lint(`how many users' items here`)).toBe(
-      `how many users' items here`
-    )
-    expect(lint(`what's going on`)).toBe(`what's going on`)
-  })
-  test('math exp', () => {
-    expect(lint('1+1=2')).toBe('1 + 1 = 2')
-    expect(lint('a|b')).toBe('a|b')
-    expect(lint('a| b')).toBe('a | b')
-    expect(lint('a |b')).toBe('a | b')
-    expect(lint('a | b')).toBe('a | b')
-    expect(lint('a||b')).toBe('a||b')
-    expect(lint('a|| b')).toBe('a || b')
-    expect(lint('a ||b')).toBe('a || b')
-    expect(lint('a || b')).toBe('a || b')
-  })
-  test('arrow chars', () => {
-    expect(lint('Chrome 顶部导航 > 窗口 > 任务管理')).toBe(
-      'Chrome 顶部导航 > 窗口 > 任务管理'
-    )
-  })
-  test('curly brackets', () => {
-    expect(lint('# 简介 {#introduction}')).toBe(
-      '# 简介 {#introduction}'
-    )
-    expect(lint('# 简介{#introduction}')).toBe(
-      '# 简介 {#introduction}'
-    )
-    expect(lint('### 托管模式 {#takeover-mode}')).toBe(
-      '### 托管模式 {#takeover-mode}'
-    )
-  })
-})
-
-describe('lint with different arguments', () => {
-  test('stirng[] argument for rules', () => {
-    expect(
-      lint('汉字和English之间需要有空格比如 half width content。', {
-        rules: ['space-full-width-content']
-      })
-    ).toBe('汉字和 English 之间需要有空格比如 half width content。')
-  })
-  test('ignored cases', () => {
-    expect(
-      lint('汉字和English之间需要有空格比如 half width content。', {
-        ignoredCases: [{ textStart: '和English之间' }]
-      })
-    ).toBe('汉字和English之间需要有空格比如 half width content。')
   })
 })
